+++ lib/foreground_service_isolate.dart
@@ -5,3 +5,4 @@ export 'src/foreground_service_isolate.dart'
 export 'src/foreground_service_type.dart';
 export 'src/isolate_utils.dart';
 export 'src/notification_details.dart';
+export 'src/notification_importance.dart';