+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.1
 
-- Adds optional `channelDescription` to `NotificationDetails`
+- Adds optional `channelDescription` and `importance` to `NotificationDetails`
 - Fixes minification issue in release builds
 
 ## 0.1.0